dmenu.c: (setup): allow dmenu to be embedded in alpha windows
for dmenu to be embedded in a window which uses
a different visual and depth, (say within st with the alpha patch),
the window needs to be created like this:

	int xv_depth = DefaultDepth(dc->dpy, screen);
	Visual * xv_visual = DefaultVisual(dc->dpy, screen);

	swa.border_pixel = normcol[ColFG];
	swa.colormap = XCreateColormap(dc->dpy, parentwin, xv_visual, AllocNone);

	win = XCreateWindow(dc->dpy, parentwin, x, y, mw, mh, 0,
			    xv_depth, CopyFromParent, xv_visual,
			    CWBorderPixel | CWColormap |
	                    CWOverrideRedirect | CWBackPixel | CWEventMask, &swa);

i.e. both the border pixel, the back pixel, AND colormap need to be
set in the attributes, where the colormap is taken from the parent.
(However the depth and visual default to the root.?!  I didn't
understand as I only expected BadMatch and create problems if the
depth and visual of the 2 windows did not map, but maybe I'm running
an old Xorg 1.10.4.)